backend/requirements.txt
Django==4.2.1
python-dotenv==1.0.0
postgres==4.0
djangorestframework==3.14.0
django-colorfield==0.8.0
webcolors==1.13
django-filter==23.2
djoser==2.2.0
psycopg2==2.9.6
drf-extra-fields==3.5.0